Fixed registration and test
