+++ passnail.core/conf/generator.properties
@@ -1,4 +1,4 @@
-#Tue Feb 23 21:43:50 CET 2021
+#Wed Feb 24 21:01:53 CET 2021
 gen.upper.case.nr=1
 gen.lower.case.nr=5
 gen.passw.length.nr=9
