+++ passnail.core/src/main/resources/application.properties
@@ -1,6 +1,5 @@
 spring.datasource.initialization-mode=always
-spring.datasource.driverClassName=org.h2.Driver
-spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
+spring.datasource.driverClassName=@value.spring.datasource.driverClassName@
+spring.jpa.database-platform=@value.spring.jpa.database-platform@
 spring.jpa.show-sql=false
 spring.jpa.generate-ddl=true
-#spring.jpa.hibernate.ddl-auto=update
