+++ passnail.core/src/test/java/com/passnail/core/tools/pass/DefaultPasswordGeneratorManagerTest.java
@@ -1,6 +1,6 @@
 package com.passnail.core.tools.pass;
 
-import com.passnail.core.main.test.ConfAttributes;
+import com.passnail.core.main.config.ConfAttributes;
 import com.passnail.core.service.gen.DefaultPasswordGeneratorManager;
 import com.passnail.core.service.gen.PasswordGeneratorManagerIf;
 import com.passnail.core.throwable.IncorrectPropertiesException;
